# >3-5x faster torch.compile forward compilation for autoregressive decoder models