… src/transformers/models/dbrx/modeling_dbrx.py
… src/transformers/models/gemma/modeling_gemma.py
… src/transformers/models/gemma2/modeling_gemma2.py
… src/transformers/models/jetmoe/modeling_jetmoe.py
… src/transformers/models/llama/modeling_llama.py
… src/transformers/models/mistral/modeling_mistral.py
… src/transformers/models/mixtral/modeling_mixtral.py
… src/transformers/models/olmo/modeling_olmo.py
… src/transformers/models/persimmon/modeling_persimmon.py
… src/transformers/models/phi/modeling_phi.py
… src/transformers/models/phi3/modeling_phi3.py
… src/transformers/models/qwen2/modeling_qwen2.py
… src/transformers/models/qwen2_moe/modeling_qwen2_moe.py
… src/transformers/models/stablelm/modeling_stablelm.py
… src/transformers/models/starcoder2/modeling_starcoder2.py
… src/transformers/models/whisper/modeling_whisper.py
… tests/models/gemma/test_modeling_gemma.py
… tests/test_modeling_common.py