As detailed in https://github.com/huggingface/transformers/pull/29114 and https://github.com/huggingface/transformers/issues/29165, the current implementation of `GenerationMixin.generate` passes 2D integer tensors of variable length as `attention_mask` to the models forward, that are transformed into 4D tensors of static shape in the models forward for decoder models.

This is an issue when using torch.compile with static cache, as then the model forward has inputs of various shape at each decoding step, which retriggers a cuda graph capture. This is especially bad when setting large enough `max_new_tokens` (>64) as cuda graphs are recaptured many many times.

Effectively as per https://pytorch.org/docs/stable/torch.compiler_cudagraph_trees.html#cudagraph-trees-integration, what is to be considered the compilation time is the `torch.compile` call, and the subsequent two forward calls (without guard invalidation, cudagraph invalidation).

For example, below, we set `max_new_tokens = 256`, with an 8B model on an A100, and on main the compilation using `model.forward = torch.compile(model.forward, mode="reduce-overhead", fullgraph=True)` takes:
```
----- GENERATE WITHOUT COMPILE
Non-compiled generate call took (no warmup): 12.546 s
decoded ['I would like (redacted)"]
compiling...
Finished compile call: 0.926 s
----- GENERATE WITH COMPILE
/home/felix/miniconda3/envs/fx/lib/python3.9/site-packages/torch/_inductor/compile_fx.py:150: UserWarning: TensorFloat32 tensor cores for float32 matrix multiplication available but not enabled. Consider setting `torch.set_float32_matmul_precision('high')` for better performance.
  warnings.warn(
First generate call took: 89.536 s (the first run warms up each graph, which does things like CuBlas or Triton benchmarking)
-------- decoded 1 ['I would like (redacted)"]
Second generate call took: 267.591 s (The second run does a CUDA Graph recording, and replays it)
-------- decoded 2 ['I would like (redacted)"]
Third generate call took: 3.421 s (Finally we hit the optimized, CUDA Graph replay path)
-------- decoded 3 ['I would like (redacted)"]

```

while with this PR the compilation takes
```
----- GENERATE WITHOUT COMPILE
Non-compiled generate call took (no warmup): 11.672 s
decoded ['I would like (redacted)"]
compiling...
Finished compile call: 0.999 s
----- GENERATE WITH COMPILE
First generate call took: 85.505 s (the first run warms up each graph, which does things like CuBlas or Triton benchmarking)
-------- decoded 1 ['I would like (redacted)"]
Second generate call took: 4.845 s (The second run does a CUDA Graph recording, and replays it)
-------- decoded 2 ['I would like (redacted)"]
Third generate call took: 3.495 s (Finally we hit the optimized, CUDA Graph replay path)
-------- decoded 3 ['I would like (redacted)"]
```

repro here https://gist.github.com/fxmarty/74b4463935ada44ef7755a544cae8773, using `dynamic=True` is even worse

cc @LysandreJik @ArthurZucker @amyeroberts @gante I think we'll need a lot of cleanup in `modeling_attn_mask_utils`  at some point - we have many `_prepare_attention_mask_xxx` functions (even some in `PreTrainedModel`)

Closes https://github.com/huggingface/transformers/issues/29165