@fxmarty 2024-07-30
> In accordance to transformers-philosophy, the causal mask creation should stay in the modeling file as it is CORE. Wrapping it in a function is not a problem !

Understood.

> This attention mask 4D created in prepare_input_for_generation should be done in generate automatically

I am not sure about this, as it is model specific (some models support `cache_position` & static cache, others don't). Do you mean that the 2D -> 4D update should preferably be handled in `_update_model_kwargs_for_generation` instead of `prepare_inputs_for_generation`?