@fxmarty 2024-07-30
I'll add some tests here.