@fxmarty 2024-07-31
I checked the slow tests. There are no other tests failing due to this PR. Although on `main` there are a few tests failing unrelated to this PR which may be concerning (on A100, maybe some reference are for an other GPU, especially gemma):

```
tests/models/cohere/test_modeling_cohere.py::CohereModelTest::test_generate_compile_fullgraph
tests/models/gemma/test_modeling_gemma.py::GemmaModelTest::test_generate_compile_fullgraph
tests/models/gemma/test_modeling_gemma.py::GemmaIntegrationTest::test_model_2b_bf16
tests/models/gemma/test_modeling_gemma.py::GemmaIntegrationTest::test_model_2b_bf16_dola
tests/models/gemma/test_modeling_gemma.py::GemmaIntegrationTest::test_model_2b_eager
tests/models/gemma/test_modeling_gemma.py::GemmaIntegrationTest::test_model_2b_sdpa
tests/models/gemma/test_modeling_gemma.py::GemmaIntegrationTest::test_model_7b_4bit
tests/models/gemma/test_modeling_gemma.py::GemmaIntegrationTest::test_model_7b_bf16
tests/models/gemma2/test_modeling_gemma2.py::GemmaModelTest::test_generate_compile_fullgraph
tests/models/gemma2/test_modeling_gemma2.py::Gemma2ModelTest::test_sdpa_can_dispatch_on_flash
tests/models/gemma2/test_modeling_gemma2.py::Gemma2ModelTest::test_sdpa_equivalence
tests/models/llama/test_modeling_llama.py::LlamaModelTest::test_generate_compile_fullgraph
tests/models/llama/test_modeling_llama.py::LlamaIntegrationTest::test_model_7b_logits
tests/models/llama/test_modeling_llama.py::LlamaIntegrationTest::test_model_7b_logits_bf16
``` 

Testing for `mistral or mixtral or olmo or persimmon or phi or phi3 or qwen2 or qwen2_moe or stablelm or starcoder2 or whisper` now.